// src/components/input/types/Types.ts
export type Input = {
  code: number,
  axis: number,
  lastAxis: number
}

export enum PointerCodes {
  LEFT = 0,
  RIGHT = 2
}

export enum KeyCodes {
  A = 65,
  B = 66,
  C = 67,
  D = 68,
  E = 69,
  F = 70,
  G = 71,
  H = 72,
  I = 73,
  J = 74,
  K = 75,
  L = 76,
  M = 77,
  N = 78,
  O = 79,
  P = 80,
  Q = 81,
  R = 82,
  S = 83,
  T = 84,
  U = 85,
  V = 86,
  W = 87,
  X = 88,
  Y = 89,
  Z = 90,
  SHIFT = 16,
  SPACE = 32,
  RETURN = 13,
  ALT = 18,
  KEYPAD_0 = 48,
  KEYPAD_1 = 49,
  KEYPAD_2 = 50,
  KEYPAD_3 = 51,
  KEYPAD_4 = 52,
  KEYPAD_5 = 53,
  KEYPAD_6 = 54,
  KEYPAD_7 = 55,
  KEYPAD_8 = 56,
  KEYPAD_9 = 57,
  COMMA = 188,
  PERIOD = 190,
  FSLASH = 191,
  QUOTE = 222,
  SEMICOLON = 186,
  OPEN_BRACKET = 219,
  CLOSE_BRACKET = 221,
  BSLASH = 220,
  TILDE = 192,
  ESCAPE = 27,
  CAPS_LOCK = 20,
  CONTROL = 17,
  WINKEY = 91,
  TAB = 9,
  F1 = 112,
  F2 = 113,
  F3 = 114,
  F4 = 115, 
  F5 = 116,
  F6 = 117,
  F7 = 118,
  F8 = 119,
  F9 = 120,
  F10 = 121,
  F11 = 122,
  F12 = 123,
  F13 = 124,
  F14 = 125,
  F15 = 126,
  F16 = 127,
  F17 = 128,
  F18 = 129,
  F19 = 130,
  DELETE = 46,
  EQUALS = 187,
  MINUS = 189,
  BACKSPACE = 8,
  INSERT = 45,
  HOME = 36,
  PAGE_UP = 33,
  PAGE_DOWN = 34,
  END = 35,
  UP_ARROW = 38,
  LEFT_ARROW = 37,
  RIGHT_ARROW = 39,
  DOWN_ARROW = 40,
  CLEAR = 144,
}